+++ src/main/java/com/hrms/repository/MedicalReimbursementDetailsDao.java
@@ -1,0 +1,8 @@
+package com.hrms.repository;
+
+import com.hrms.dao.GenericDao;
+import com.hrms.model.MedicalReimbursementDetail;
+
+public interface MedicalReimbursementDetailsDao extends GenericDao<MedicalReimbursementDetail> {
+
+}
