+++ src/main/java/com/hrms/repository/MedicalReimbursementDetailsDaoImpl.java
@@ -1,0 +1,11 @@
+package com.hrms.repository;
+
+import org.springframework.stereotype.Repository;
+
+import com.hrms.dao.AbstractGenericDao;
+import com.hrms.model.MedicalReimbursementDetail;
+
+@Repository
+public class MedicalReimbursementDetailsDaoImpl extends AbstractGenericDao<MedicalReimbursementDetail> implements MedicalReimbursementDetailsDao {
+
+}
