+++ src/main/java/com/hrms/service/MedicalReimbursementDetailsService.java
@@ -1,0 +1,9 @@
+package com.hrms.service;
+
+import com.hrms.model.MedicalReimbursementDetail;
+
+public interface MedicalReimbursementDetailsService {
+
+	boolean addMedicalReimbursementDetails(MedicalReimbursementDetail m4);
+
+}
