+++ src/main/java/com/hrms/service/MedicalReimbursementServiceImpl.java
@@ -15,7 +15,7 @@ MedicalReimbursementDao medicalReimbursementDao;
 
 @Override
 public boolean addMedicalReimbursement(MedicalReimbursement m2) {
-	m2.setInsDate(new Date());
+	
 	m2.setSlipNo(medicalReimbursementDao.getMaxId("SLP"));
 	this.medicalReimbursementDao.saveOrUpdate(m2);
 	return true;
